#feat. 거리응답칼럼을 BigDecimal로 타입변경 후 소수점 4자리까지만 보이고 반올림하도록 수정

+++ src/main/java/kr/co/kbhealthcareproject/dto/response/DailyActivityResponseDto.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 public class DailyActivityResponseDto implements Serializable {
@@ -14,11 +15,11 @@ public class DailyActivityResponseDto implements Serializable {
     @JsonProperty(value = "calories")
     private int totalCalories;
     @JsonProperty(value = "distance")
-    private double totalDistance;
+    private BigDecimal totalDistance;
     @JsonProperty(value = "recordkey")
     private String recordKey;
 
-    public DailyActivityResponseDto(String daily, int totalSteps, int totalCalories, double totalDistance, String recordKey) {
+    public DailyActivityResponseDto(String daily, int totalSteps, int totalCalories, BigDecimal totalDistance, String recordKey) {
         this.daily = daily;
         this.totalSteps = totalSteps;
         this.totalCalories = totalCalories;
